+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - When the unix path is bigger than what can fit into `struct sockaddr_un`, a shorter temporary path will be used instead and a symlink created at the original path.
+
 ## v0.17.1
 
   - Add `#to_s` and `#inspect` for `IO::Endpoint::NamedEndpoints`.